[IMP] business_requirement_deliverable: Auto-subscribe to BR followers
When creating a business requirement deliverable (BRD), auto-subscribe followers of the
parent business requirement (BR), similar to what is done on tasks/projects.

This is performed automatically thanks to mail mechanisms, having a mail.message.subtype
in BR having other subtype in BRD as parent. We put them as hidden, as that subtypes
don't have other function than that.